+++ .github/workflows/eslint.yml
@@ -12,10 +12,10 @@ jobs:
         with:
           version: 7
 
-      - name: Node.JS 16
+      - name: Node.JS 18
         uses: actions/setup-node@v3
         with:
-          node-version: 16
+          node-version: 18
           cache: 'pnpm'
