--- app/(DynamicLayout)/projects/[slug]/head.tsx
@@ -1,71 +1,0 @@
-import PayloadResponse from 'types/PayloadResponse';
-import { Media, Project } from 'types/payload-types';
-import Head from 'ui/Head';
-import { getImageUrl } from 'ui/Image';
-
-interface ProjectData {
-	en: {
-		title: string;
-		shortDescription: string;
-		slug: string;
-		image: Media;
-		ogImage?: Media;
-	};
-	jp: {
-		title?: string;
-		shortDescription?: string;
-	}
-}
-
-async function fetchProject(slug: string): Promise<ProjectData | null> {
-	// Fetch EN and JP version for page, CMS will fallback to EN for any fields not translated
-	const enProjectRes = await fetch(`${process.env.NEXT_PUBLIC_CMS_URL!}/api/projects?where[slug][like]=${slug}&depth=2`, {
-		headers: {
-			'X-RateLimit-Bypass': process.env.PAYLOAD_BYPASS_RATE_LIMIT_KEY ?? '',
-		} as Record<string, string>,
-	});
-	const parsedEnProjectRes = (await enProjectRes.json() as PayloadResponse<Project>);
-	if (parsedEnProjectRes.totalDocs === 0) return null;
-	const enProject = parsedEnProjectRes.docs[0];
-
-	const jpProjectRes = await fetch(`${process.env.NEXT_PUBLIC_CMS_URL!}/api/projects?where[slug][like]=${slug}&depth=0&locale=jp`, {
-		headers: {
-			'X-RateLimit-Bypass': process.env.PAYLOAD_BYPASS_RATE_LIMIT_KEY ?? '',
-		} as Record<string, string>,
-	});
-	const jpProject = (await jpProjectRes.json() as PayloadResponse<Project>).docs[0];
-
-	return {
-		en: {
-			title: enProject.title,
-			shortDescription: enProject.shortDescription,
-			slug: enProject.slug,
-			image: enProject.image as Media,
-			ogImage: enProject.ogImage as Media,
-		},
-		jp: {
-			title: jpProject.title ?? null,
-			shortDescription: jpProject.shortDescription ?? null,
-		},
-	};
-}
-
-export default async function PageHead({ params }: { params: { slug: string } }) {
-	const project = await fetchProject(params.slug);
-
-	return (
-		<Head
-			/* TODO: Make customizable? */
-			color="#FF3D3D"
-			/* TODO: Use locale */
-			title={project?.en.title ?? undefined}
-			description={project?.en.shortDescription ?? undefined}
-			keywords={['guratanabata']}
-			image={
-				(project
-					? getImageUrl({ src: project.en.ogImage?.url ?? project.en.image.url!, width: 1024 })
-					: 'https://holoen.fans/img/logo.png')
-			}
-		/>
-	);
-}
