--- app/(StaticLayout)/head.tsx
@@ -1,7 +1,0 @@
-import Head from 'ui/Head';
-
-export default function PageHead() {
-	return (
-		<Head />
-	);
-}
