--- app/(StaticLayout)/projects/head.tsx
@@ -1,10 +1,0 @@
-import Head from 'ui/Head';
-
-export default function PageHead() {
-	return (
-		<Head
-			title="HoloEN Fan Website - Projects"
-			description="A list of all the projects organized by Hololive EN Fan servers!"
-		/>
-	);
-}
