+++ app/layout.tsx
@@ -1,5 +1,6 @@
 import 'styles/globals.css';
 import DarkModeContextProvider from 'ui/DarkModeContextProvider';
+import { Metadata } from 'next';
 
 interface IProps {
 	children: React.ReactNode;
@@ -14,3 +15,30 @@ export default async function RootLayout({ children }: IProps) {
 		</html>
 	);
 }
+
+export const metadata: Metadata = {
+	title: 'HoloEN Fan Website',
+	description: 'A website featuring amazing projects organized by the community!',
+	keywords: ['hololive en', 'hef', 'hololive fan', 'hololive en fan', 'hololive'],
+	themeColor: '#FF3D3D',
+	openGraph: {
+		title: 'HoloEN Fan Website',
+		description: 'A website featuring amazing projects organized by the community!',
+		images: 'https://holoen.fans/img/logo.png',
+		type: 'website',
+		siteName: 'HoloEN Fan Website',
+	},
+	twitter: {
+		title: 'HoloEN Fan Website',
+		description: 'A website featuring amazing projects organized by the community!',
+		images: 'https://holoen.fans/img/logo.png',
+		site: '@HEF_Website',
+		creator: '@GoldElysium',
+		card: 'summary_large_image',
+	},
+	other: {
+		subject: 'HoloEN Fan Website',
+		language: 'EN',
+		url: 'https://holoen.fans',
+	},
+};
