+++ package.json
@@ -20,54 +20,54 @@
     "pnpm": ">=7"
   },
   "dependencies": {
-    "@emotion/react": "11.10.5",
-    "@emotion/styled": "11.10.5",
-    "@headlessui/react": "1.7.7",
-    "@heroicons/react": "2.0.13",
+    "@emotion/react": "11.10.6",
+    "@emotion/styled": "11.10.6",
+    "@headlessui/react": "1.7.13",
+    "@heroicons/react": "2.0.16",
     "@mdi/js": "^7.1.96",
     "@mdi/react": "^1.6.1",
-    "daisyui": "^2.50.1",
+    "daisyui": "^2.51.3",
     "date-fns": "2.29.3",
     "escape-html": "^1.0.3",
-    "nanoid": "4.0.0",
-    "next": "13.1.4",
+    "nanoid": "4.0.1",
+    "next": "13.2.3",
     "phaser": "3.55.2",
     "phaser3-nineslice": "0.5.0",
-    "phaser3-rex-plugins": "1.1.79",
+    "phaser3-rex-plugins": "1.1.82",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-infinite-scroll-component": "6.1.0",
-    "react-player": "2.11.0",
+    "react-player": "2.11.2",
     "server-only": "^0.0.1",
-    "slate": "^0.88.1",
-    "styled-components": "5.3.6",
-    "swr": "^2.0.3",
+    "slate": "^0.91.4",
+    "styled-components": "5.3.8",
+    "swr": "^2.0.4",
     "webfontloader": "1.6.28"
   },
   "devDependencies": {
-    "@commitlint/cli": "17.4.2",
-    "@commitlint/config-conventional": "17.4.2",
-    "@next/eslint-plugin-next": "^13.1.6",
+    "@commitlint/cli": "17.4.4",
+    "@commitlint/config-conventional": "17.4.4",
+    "@next/eslint-plugin-next": "^13.2.3",
     "@types/escape-html": "^1.0.2",
-    "@types/react": "18.0.27",
+    "@types/react": "18.0.28",
     "@types/safe-json-stringify": "1.1.2",
     "@types/webfontloader": "1.6.35",
-    "@typescript-eslint/eslint-plugin": "5.48.2",
-    "@typescript-eslint/parser": "5.48.2",
+    "@typescript-eslint/eslint-plugin": "5.54.0",
+    "@typescript-eslint/parser": "5.54.0",
     "autoprefixer": "10.4.13",
-    "eslint": "8.32.0",
+    "eslint": "8.35.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-airbnb-typescript": "17.0.0",
-    "eslint-config-next": "13.1.4",
+    "eslint-config-next": "13.2.3",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-jsx-a11y": "6.7.1",
-    "eslint-plugin-react": "7.32.1",
+    "eslint-plugin-react": "7.32.2",
     "eslint-plugin-react-hooks": "4.6.0",
     "husky": "8.0.3",
-    "lint-staged": "13.1.0",
+    "lint-staged": "13.1.2",
     "postcss": "8.4.21",
-    "tailwindcss": "3.2.4",
-    "typescript": "4.9.4"
+    "tailwindcss": "3.2.7",
+    "typescript": "4.9.5"
   },
   "overrides": {
     "prosemirror-tables": "^1.1.1"
